+++ src/codegen.cpp
@@ -132,7 +132,11 @@ void ByteCodeGenerator::emitMethodCallIns(unsigned int paramSize, const MethodHa
   if (handle.isNative()) {
     this->emitNativeCallIns(paramSize + 1, handle.getIndex(), !handle.getReturnType().isVoidType());
   } else {
-    fatal("unsupported\n");
+    this->emitValIns(OpCode::CALL_METHOD, paramSize, 0);
+    this->curBuilder().append16(handle.getIndex());
+    if (handle.getReturnType().isVoidType()) {
+      this->emit0byteIns(OpCode::POP);
+    }
   }
 }
 
@@ -1607,7 +1611,7 @@ void ByteCodeDumper::dumpCode(const ydsh::CompiledCode &c) {
         fprintf(this->fp, "  %s", this->typePool.get(v).getName());
       } else {
         const int byteSize = getByteSize(code);
-        if (code == OpCode::FORK) {
+        if (code == OpCode::FORK || code == OpCode::CALL_METHOD) {
           fprintf(this->fp, "  %d  %d", read8(c.getCode(), i + 1), read16(c.getCode(), i + 2));
         } else if (code == OpCode::RECLAIM_LOCAL || code == OpCode::ADD_GLOBBING ||
                    code == OpCode::INIT_FIELDS) {